python/langgraph/agents/arxiv-researcher/requirements.txt
python-dotenv
typing-extensions
langgraph
langgraph-supervisor
langchain
langchain-core
langchain-openai
langchain-anthropic
langgraph-supervisor
langchain-community
scikit-learn
openai
ipython
langgraph-sdk
langgraph-api
langchain-chroma
beautifulsoup4
ruff
langgraph-cli
langgraph-runtime-inmem
pydantic
langchain-mcp-adapters
supabase
aiohttp
langchain-anthropic
langgraph-cli[inmem]
PyMuPDF
ipykernel